refactor(civitai): extract parse_civitai_url
Move parse_civitai_url() from lora.py and embedding.py to civitai.py,
eliminating ~60 lines of duplicate code. Update all import sites and
consolidate duplicate test class.
