+++ src/oneiro/pipelines/embedding.py
@@ -1,12 +1,11 @@
 """Textual inversion / embedding configuration types and loading utilities."""
 
-import re
 from dataclasses import dataclass, field
 from enum import Enum
 from pathlib import Path
 from typing import Any
 
-from oneiro.civitai import CivitaiClient
+from oneiro.civitai import CivitaiClient, parse_civitai_url
 from oneiro.pipelines.lora import PIPELINE_BASE_MODEL_MAP
 
 
@@ -67,38 +66,6 @@ class EmbeddingConfig:
         elif self.source == EmbeddingSource.LOCAL:
             if not self.path:
                 raise ValueError("local source requires path")
-
-
-def parse_civitai_url(url: str) -> tuple[int, int | None]:
-    """Parse Civitai URL to extract model ID and optional version ID.
-
-    Supports formats:
-    - https://civitai.com/models/12345
-    - https://civitai.com/models/12345/model-name
-    - https://civitai.com/models/12345?modelVersionId=67890
-    - https://civitai.com/models/12345/name?modelVersionId=67890
-
-    Args:
-        url: Civitai model URL
-
-    Returns:
-        Tuple of (model_id, version_id or None)
-
-    Raises:
-        ValueError: If URL format is invalid
-    """
-    # Match model ID in path
-    model_match = re.search(r"/models/(\d+)", url)
-    if not model_match:
-        raise ValueError(f"Invalid Civitai URL format: {url}")
-
-    model_id = int(model_match.group(1))
-
-    # Check for version in query string
-    version_match = re.search(r"modelVersionId=(\d+)", url)
-    version_id = int(version_match.group(1)) if version_match else None
-
-    return model_id, version_id
 
 
 def parse_embedding_config(
